Provide better error reporting from the Amazon::S3 module
Case CPANEL-1618

Some customers, transporting backups across an unreliable
connection, were getting the following error:
undefined value as an ARRAY reference at
/usr/local/cpanel/3rdparty/perl/514/lib64/perl5/cpanel_lib/Amazon/S3.pm.in
line
212.
This was due to the server not responding and the
subsequent code attempting to de-reference a null response.
I've put checks for a null reponse and throw a more
meaningful error.