Improve error reporting for XML parse exceptions
Case CPANEL-13904,HB-3261:

If XML::Simple encounters an error parsing the XML returned
from Amazon we get a very terse error like, "no element found"
which tells us nothing.  In the event that we encounter this,
we at least need a copy of the XML it failed to parse as well
as the stack trace so we know what operation it was trying
to perform.