Allow Amazon::S3 to handle bucket names containing periods
Case CPANEL-18897,HB-3713:

When we check the region for the bucket we perform an
http request on <bucket name>.<amazon host>.  Amazon's
response is signed by a wild-card certificate
(*.s3.amazonaws.com).  Normally this works, but if the
bucket name contains a period, then the certificate
check fails since it is expecting a certificate for
a subdomain.  So, for the region check, we need to
set it to expect the certificate for Amazon's host.