Switch from deprecated path-style URLs when possible
Case CPANEL-31814:  Amazon has deprecated the use of
path-style URLs, i.e. where the URL path begins bucket name,
in favor of virtual-hosted style where the bucket name
is a subdomain of the host:
https://aws.amazon.com/blogs/aws/amazon-s3-path-deprecation-plan-the-rest-of-the-story/
I've modified it to use the virtual-style where possible,
where the bucket name can be a valid DNS segment.
I still use the path-style for bucket names with periods
in them.  As per the link, this is what Amazon's SDK currently
does as these would currently cause SSL certificate errors.
Amazon plans on addressing this before the cutoff date.